requirements.txt
# Core Data Science Libraries
numpy>=1.26.0
pandas>=2.1.0
scipy>=1.11.0

# Machine Learning & Statistical Models
scikit-learn>=1.3.0
statsmodels>=0.14.0
xgboost>=2.0.0

# Time Series Analysis
pmdarima>=2.0.4

# Data Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0

# Data Processing
openpyxl>=3.1.2

# Model Persistence
joblib>=1.3.2

# Configuration
pyyaml>=6.0.1

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
